+++ packages/tribe-client/src/cli.ts
@@ -75,3 +75,5 @@ async function main(): Promise<void> {
 }
 
 await main()
+
+export {}